python/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
chromadb>=0.5.0